% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCity.R
\name{plotCity}
\alias{plotCity}
\title{Visualizatio of all top cities.}
\value{
Return ggplot of all top cities.
}
\description{
Visualization of all top cities from shiny app.
}
\examples{

\dontshow{
library(ppcSpatial)
plotCity()
}

\dontrun{Launch_ppcSpatial()}
}
\keyword{internal}
