% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistricts.R
\name{plotDistricts}
\alias{plotDistricts}
\title{Visualization of selected districts.}
\arguments{
\item{filterTehsil}{The selected districts for which plot to be build.}
}
\value{
Return ggplot of selected districts.
}
\description{
Visualization of selected districts from shiny app.
}
\keyword{internal}
