% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDivisions.R
\name{plotDivisions}
\alias{plotDivisions}
\title{Visualization of selected divisions.}
\arguments{
\item{filterTehsil}{The selected divisions for which plot to be build.}
}
\value{
Return ggplot of selected divisions.
}
\description{
Visualization of selected divisions from shiny app.
}
\keyword{internal}
