% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProvinces.R
\name{plotProvinces}
\alias{plotProvinces}
\title{Visualization of all provinces.}
\value{
Return ggplot of all provinces.
}
\description{
Visualization of all provinces from shiny app.
}
\keyword{internal}
