% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTehsils.R
\name{plotTehsils}
\alias{plotTehsils}
\title{Visualization of selected tehsils.}
\arguments{
\item{filterTehsil}{The selected tehsils for which plot to be build.}
}
\value{
Return ggplot of selected tehsils.
}
\description{
Visualization of selected tehsils from shiny app.
}
\keyword{internal}
